Fixes for bug 3062102: 2.0 regressed and failed to serialize maps containing null for keys, or values.
Fixes bug for deserializing Timestamp and subclasses of Date that don't expose a no-arg constructor.  Now we will instantiate it directly using the new SomeDateObject(long) method.  If that fails to find we'll use the no-arg constructor.

Reduced setup code and reused the FixtureCreator code.

Finally added a new ObjectFactory to deserialize into an existing object as well as extra handy methods for doing so without having to specify the ObjectFactory directly.